+++ resources/views/components/hero.blade.php
@@ -31,43 +31,65 @@
     'layout' => Config::getArray('hyde-layouts-manager.components.hero.default.layout', []),
 ])
 
-<!-- Render Hero -->
-<div class="{{ $settings['bgColor'] }} {{ $settings['darkBgColor'] }} {{ $settings['padding'] }}">
-    <div class="py-8 px-4 mx-auto max-w-screen-xl text-{{ $settings['align'] }} lg:py-16">
+<!-- Render Hero component (Jumbotron) -->
+<section class="{{ $settings['bgColor'] }} {{ $settings['darkBgColor'] }}">
+    <div class="{{ $settings['padding'] }} mx-auto {{ $settings['maxWidth'] }} lg:py-16">
 
-        <!-- Hero Heading (always visible if provided) -->
-        <h1 class="mb-4 text-4xl font-extrabold tracking-tight leading-none text-gray-900 md:text-5xl lg:text-6xl dark:text-white">
+        <!-- Heading -->
+        <h1 class="mb-4
+            {{ $settings['headingTextSize']['default'] }}
+            md:{{ $settings['headingTextSize']['md'] }}
+            lg:{{ $settings['headingTextSize']['lg'] }}
+            font-extrabold tracking-tight leading-none
+            text-gray-900 dark:text-white text-{{ $settings['headingTextAlign'] }}">
             {{ $settings['headingText'] }}
         </h1>
 
-        <!-- Hero Sub Heading (conditionally rendered based on showSubHeadingText) -->
-        @if (isset($layout['showSubHeadingText']) && $layout['showSubHeadingText'])
-            <p class="mb-8 text-lg font-normal text-gray-500 lg:text-xl sm:px-16 lg:px-48 dark:text-gray-400">
-                {{ $settings['subHeadingText'] }}
-            </p>
-        @endif
+        <!-- Subheading -->
+        <p class="mb-8
+            {{ $settings['subHeadingTextSize']['default'] }}
+            md:{{ $settings['subHeadingTextSize']['md'] }}
+            font-normal text-gray-500 lg:text-xl dark:text-gray-400
+            text-{{ $settings['subHeadingTextAlign'] }}">
+            {{ $settings['subHeadingText'] }}
+        </p>
 
         <!-- Button Group -->
-        <div class="flex flex-col space-y-4 sm:flex-row sm:justify-center sm:space-y-0">
-            <!-- Primary Button (conditionally rendered based on showPrimaryButton) -->
-            @if (isset($layout['showPrimaryButton']) && $layout['showPrimaryButton'] && isset($settings['primaryButton']['link']))
-                <a href="{{ $settings['primaryButton']['link'] ?? '#' }}"
-                   class="inline-flex justify-center items-center gap-2 py-3 px-5 text-base font-medium text-center {{ $settings['primaryButton']['bgColor'] }} {{ $settings['primaryButton']['darkBgColor'] }} {{ $settings['primaryButton']['textColor'] }} rounded-lg hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 dark:focus:ring-blue-900">
-                    {!! $settings['primaryButton']['text'] !!}
+        <div class="flex flex-col space-y-4 sm:flex-row justify-{{
+            $settings['buttonsGroupAlign'] === 'left' ? 'start' :
+            ($settings['buttonsGroupAlign'] === 'right' ? 'end' : 'center') }} sm:space-y-0 w-full">
+            <!-- Primary Button -->
+            @if ($layout['showPrimaryButton'] && isset($settings['primaryButton']['link']))
+                <button type="button" class="inline-flex justify-center items-center py-3 px-5 {{ $settings['buttonTextSize']['default'] }}
+                   sm:{{ $settings['buttonTextSize']['sm'] }} font-medium text-center
+                   {{ $settings['primaryButton']['textColor'] }}
+                   rounded-lg {{ $settings['primaryButton']['bgColor'] }}
+                   {{ $settings['primaryButton']['hoverBgColor'] }} focus:ring-4
+                   {{ $settings['primaryButton']['focusRingColor'] }}
+                   dark:focus:ring-{{ $settings['primaryButton']['darkFocusRingColor'] }} w-auto">
+                    <a href="{{ $settings['primaryButton']['link'] }}">
+                        {!! $settings['primaryButton']['text'] !!}
+                    </a>
                     <svg class="w-3.5 h-3.5 ms-2 rtl:rotate-180" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 14 10">
                         <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M1 5h12m0 0L9 1m4 4L9 9"/>
                     </svg>
-                </a>
+                </button>
             @endif
 
-            <!-- Secondary Button (conditionally rendered based on showSecondaryButton) -->
-            @if (isset($layout['showSecondaryButton']) && $layout['showSecondaryButton'] && isset($settings['secondaryButton']['link']))
-                <a href="{{ $settings['secondaryButton']['link'] ?? '#' }}"
-                   class="py-3 px-5 sm:ms-4 text-sm font-medium text-gray-900 focus:outline-none {{ $settings['secondaryButton']['bgColor'] }} {{ $settings['secondaryButton']['textColor'] }} rounded-lg border border-gray-200 hover:bg-gray-100 hover:text-blue-700 focus:z-10 focus:ring-4 focus:ring-gray-100 dark:focus:ring-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:border-gray-600 dark:hover:text-white dark:hover:bg-gray-700">
-                    {{ $settings['secondaryButton']['text'] }}
-                </a>
+            <!-- Secondary Button -->
+            @if ($layout['showSecondaryButton'] && isset($settings['secondaryButton']['link']))
+                <button type="button" class="py-3 px-5 sm:ms-4 text-sm font-medium {{ $settings['secondaryButton']['textColor'] }} focus:outline-none
+                   {{ $settings['secondaryButton']['bgColor'] }} rounded-lg border
+                   {{ $settings['secondaryButton']['borderColor'] }}
+                   {{ $settings['secondaryButton']['hoverBgColor'] }} focus:z-10
+                   focus:ring-4 {{ $settings['secondaryButton']['focusRingColor'] }}
+                   dark:focus:ring-{{ $settings['secondaryButton']['darkFocusRingColor'] }} w-auto">
+                    <a href="{{ $settings['secondaryButton']['link'] }}">
+                        {{ $settings['secondaryButton']['text'] }}
+                    </a>
+                </button>
             @endif
         </div>
 
     </div>
-</div>
+</section>