+++ resources/views/layouts/melasistema/posts/article-excerpt.blade.php
@@ -7,9 +7,10 @@
     @endif
 
     <a href="{{ $post->getRoute() }}" class="flex flex-col items-center bg-white border border-gray-200 rounded-lg shadow md:flex-row hover:bg-gray-100 dark:border-gray-700 dark:bg-gray-800 dark:hover:bg-gray-700">
-        {{--<img class="object-cover w-full rounded-t-lg h-96 md:h-auto md:w-48 md:rounded-none md:rounded-s-lg" src="/docs/images/blog/image-4.jpg" alt="">--}}
-        @includeWhen(isset($post->image), 'hyde-layouts-manager::layouts.melasistema.posts.post.image')
-        <div class="flex flex-col justify-between p-4 leading-normal">
+        <div>
+            @includeWhen(isset($post->image), 'hyde-layouts-manager::layouts.melasistema.posts.post.image')
+        </div>
+        <div class="flex flex-col justify-between p-4 leading-normal w-3/4 md:w-2/3">
             <h5 class="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">{{ $post->data('title') ?? $post->title }}</h5>
             <div class="mb-8 text-xs">
                 @isset($post->date)
@@ -31,5 +32,4 @@
             @endisset
         </div>
     </a>
-
 </article>