+++ tailwind-layouts-manager.config.js
@@ -2,13 +2,14 @@
  * This is the Tailwind CSS configuration file for the HydePHP Layouts Manager package.
  *
  * @package Melasistema\HydeLayoutsManager
+ * @author  Luca Visciola
  * @copyright 2024 Luca Visciola
  * @license MIT License
  *
  * To include this configuration in your project's `tailwind.config.js`,
  * you need to require and merge it with your existing configuration or
  * use the command from the package `tailwind:merge`.
- * @see [README.md#installation] for more information on how to install the package. * @author  Luca Visciola
+ * @see [README.md#installation](https://github.com/melasistema/hydephp-layouts-manager/blob/master/README.md#installation)
  *
  **/
 const fs = require('fs');
@@ -24,8 +25,10 @@ const fontConfigPath = './config/hyde-layouts-manager-fonts.json';
 const userFonts = fs.existsSync(fontConfigPath)
     ? JSON.parse(fs.readFileSync(fontConfigPath, 'utf8'))
     : {};
-const melasistemaFonts = userFonts?.layouts?.melasistema?.families || {};
-const typographyMapping = userFonts?.layouts?.melasistema?.typography_mapping || {};
+
+// Dynamically load font configurations based on the layout specified in the .env DEFAULT_LAYOUT
+const layoutFonts = userFonts?.layouts?.[process.env.DEFAULT_LAYOUT || 'melasistema']?.families || {};
+const typographyMapping = userFonts?.layouts?.[process.env.DEFAULT_LAYOUT || 'melasistema']?.typography_mapping || {};
 
 module.exports = {
     darkMode: 'class',
@@ -37,35 +40,37 @@ module.exports = {
     theme: {
         extend: {
             fontFamily: {
-                sans: [safeFontSplit(melasistemaFonts.primary), 'system-ui', 'sans-serif'],
-                primary: [safeFontSplit(melasistemaFonts.primary), 'system-ui', 'sans-serif'],
-                secondary: [safeFontSplit(melasistemaFonts.secondary), 'system-ui', 'sans-serif'],
-                display: [safeFontSplit(melasistemaFonts.display), 'serif'],
-                heading: [safeFontSplit(melasistemaFonts.heading), 'sans-serif'],
-                subheading: [safeFontSplit(melasistemaFonts.subheading), 'sans-serif'],
-                accent: [safeFontSplit(melasistemaFonts.accent), 'cursive'],
-                code: [safeFontSplit(melasistemaFonts.code), 'monospace'],
-                small: [safeFontSplit(melasistemaFonts.small), 'sans-serif'],
+                sans: [safeFontSplit(layoutFonts.primary), 'system-ui', 'sans-serif'],
+                primary: [safeFontSplit(layoutFonts.primary), 'system-ui', 'sans-serif'],
+                secondary: [safeFontSplit(layoutFonts.secondary), 'system-ui', 'sans-serif'],
+                display: [safeFontSplit(layoutFonts.display), 'serif'],
+                heading: [safeFontSplit(layoutFonts.heading), 'sans-serif'],
+                subheading: [safeFontSplit(layoutFonts.subheading), 'sans-serif'],
+                accent: [safeFontSplit(layoutFonts.accent), 'cursive'],
+                code: [safeFontSplit(layoutFonts.code), 'monospace'],
+                small: [safeFontSplit(layoutFonts.small), 'sans-serif'],
             },
         },
     },
     plugins: [
         require('flowbite'),
         plugin(function ({ addBase }) {
-            addBase({
-                'h1': { fontFamily: safeFontSplit(melasistemaFonts[typographyMapping.h1] || 'sans') },
-                'h2': { fontFamily: safeFontSplit(melasistemaFonts[typographyMapping.h2] || 'sans') },
-                'h3': { fontFamily: safeFontSplit(melasistemaFonts[typographyMapping.h3] || 'subheading') },
-                'h4': { fontFamily: safeFontSplit(melasistemaFonts[typographyMapping.h4] || 'subheading') },
-                'h5': { fontFamily: safeFontSplit(melasistemaFonts[typographyMapping.h5] || 'secondary') },
-                'h6': { fontFamily: safeFontSplit(melasistemaFonts[typographyMapping.h6] || 'secondary') },
-                'p': { fontFamily: safeFontSplit(melasistemaFonts[typographyMapping.p] || 'sans') },
-                'small': { fontFamily: safeFontSplit(melasistemaFonts[typographyMapping.small] || 'small') },
-                'code': { fontFamily: safeFontSplit(melasistemaFonts[typographyMapping.code] || 'code') },
-                'blockquote': { fontFamily: safeFontSplit(melasistemaFonts[typographyMapping.blockquote] || 'display') },
-                'label': { fontFamily: safeFontSplit(melasistemaFonts[typographyMapping.label] || 'secondary') },
-                'button': { fontFamily: safeFontSplit(melasistemaFonts[typographyMapping.button] || 'heading') }
-            });
+            if (process.env.DEFAULT_LAYOUT !== 'hyde') {
+                addBase({
+                    'h1': { fontFamily: safeFontSplit(layoutFonts[typographyMapping.h1] || 'system-ui') },
+                    'h2': { fontFamily: safeFontSplit(layoutFonts[typographyMapping.h2] || 'system-ui') },
+                    'h3': { fontFamily: safeFontSplit(layoutFonts[typographyMapping.h3] || 'system-ui') },
+                    'h4': { fontFamily: safeFontSplit(layoutFonts[typographyMapping.h4] || 'system-ui') },
+                    'h5': { fontFamily: safeFontSplit(layoutFonts[typographyMapping.h5] || 'system-ui') },
+                    'h6': { fontFamily: safeFontSplit(layoutFonts[typographyMapping.h6] || 'system-ui') },
+                    'p': { fontFamily: safeFontSplit(layoutFonts[typographyMapping.p] || 'system-ui') },
+                    'small': { fontFamily: safeFontSplit(layoutFonts[typographyMapping.small] || 'sans-serif') },
+                    'code': { fontFamily: safeFontSplit(layoutFonts[typographyMapping.code] || 'monospace') },
+                    'blockquote': { fontFamily: safeFontSplit(layoutFonts[typographyMapping.blockquote] || 'serif') },
+                    'label': { fontFamily: safeFontSplit(layoutFonts[typographyMapping.label] || 'system-ui') },
+                    'button': { fontFamily: safeFontSplit(layoutFonts[typographyMapping.button] || 'system-ui') },
+                });
+            }
         }),
     ],
 };